Refactor part II, moves all support functions out of Robot Player - SJD
